Handle invalid config
